feat: Add HomeSlideshow component and update page layout
This commit adds the HomeSlideshow component to the project, which is used to display a slideshow of images on the home page. The page layout is also updated to use the HomeSlideshow component instead of the Icons.logoIcon component. This change improves the visual appeal of the home page and provides a better user experience.